12 Initial commit of UI
The UI was implemented based on previous code already created PRs for.
In order to get this on repo as well, this initial commit is needed.
If applied, the newly created Tk based form is shown using new parameters
"-u"/"--show_viewer". Also origin now points to this branch to get click and link functionallity tested
In order to get useable results, the new param -desc/--descending was required. Its a switch between git log --date-order (False) and --reverse (True)