39 Fix issue with command line argument being un-escaped
Prior this change, GitObserver called external program like "--since=1 week ago" which is not working when white space is interpreted as next argument
If applied, the colletion of log information is changed to use IO streams of stdout or file byte stream when testing instead.
The whole text block is not transported anywhere by this, since a parsing of commits can be done line by line.
Additionally the new logging functionallity is used now beside making test flag a instance member