Issue #50 Fixing an unclosed file from the logger.py If applied, this PR fixes an `ResourceWarning`, due an unclosed file, handled by the `logging` module.
After some research this behavior is "intended", but breaks our unit test, which deletes temporary log files.

Now the class `Logger` contains a method, that closes any active instances from the file handler.
The usage of `close_logger()` is only intended for unit tests, not for production. Since it would result in a loss of the logging buffer.

Also uncommented the logger unit test and adapted it with the new, mentioned, method.

Signed-off-by: worstprgr <adam@seishin.io>